+++ Allay-API/src/main/java/cn/allay/api/math/vector/ImplMutableLoc3d.java
@@ -60,7 +60,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
     }
 
     public ImplMutableLoc3d(final float x, final float y, final float z, final double pitch, final double yaw, final double headYaw, final World world) {
-        this((double) x, (double) y, (double) z, pitch, yaw, headYaw, world);
+        this(x, y, (double) z, pitch, yaw, headYaw, world);
     }
 
     public ImplMutableLoc3d(final double x, final double y, final double z, final double pitch, final double yaw, final double headYaw, final World world) {
@@ -206,7 +206,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d add(final float x, final float y, final float z) {
-        return this.add((double) x, (double) y, (double) z);
+        return this.add(x, y, (double) z);
     }
 
     @Override
@@ -221,7 +221,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d sub(final float x, final float y, final float z) {
-        return this.sub((double) x, (double) y, (double) z);
+        return this.sub(x, y, (double) z);
     }
 
     @Override
@@ -246,7 +246,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d mul(final float x, final float y, final float z) {
-        return this.mul((double) x, (double) y, (double) z);
+        return this.mul(x, y, (double) z);
     }
 
     @Override
@@ -271,7 +271,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d div(final float x, final float y, final float z) {
-        return this.div((double) x, (double) y, (double) z);
+        return this.div(x, y, (double) z);
     }
 
     @Override
@@ -286,7 +286,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public double dot(final float x, final float y, final float z) {
-        return this.dot((double) x, (double) y, (double) z);
+        return this.dot(x, y, (double) z);
     }
 
     @Override
@@ -301,7 +301,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d project(final float x, final float y, final float z) {
-        return this.project((double) x, (double) y, (double) z);
+        return this.project(x, y, (double) z);
     }
 
     @Override
@@ -321,7 +321,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d cross(final float x, final float y, final float z) {
-        return this.cross((double) x, (double) y, (double) z);
+        return this.cross(x, y, (double) z);
     }
 
     @Override
@@ -371,7 +371,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d min(final float x, final float y, final float z) {
-        return this.min((double) x, (double) y, (double) z);
+        return this.min(x, y, (double) z);
     }
 
     @Override
@@ -386,7 +386,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public MutableLoc3d max(final float x, final float y, final float z) {
-        return this.max((double) x, (double) y, (double) z);
+        return this.max(x, y, (double) z);
     }
 
     @Override
@@ -401,7 +401,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public double distanceSquared(final float x, final float y, final float z) {
-        return this.distanceSquared((double) x, (double) y, (double) z);
+        return this.distanceSquared(x, y, (double) z);
     }
 
     @Override
@@ -419,12 +419,12 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public double distance(final float x, final float y, final float z) {
-        return this.distance((double) x, (double) y, (double) z);
+        return this.distance(x, y, (double) z);
     }
 
     @Override
     public double distance(final double x, final double y, final double z) {
-        return (double) Math.sqrt(this.distanceSquared(x, y, z));
+        return Math.sqrt(this.distanceSquared(x, y, z));
     }
 
     @Override
@@ -434,7 +434,7 @@ final class ImplMutableLoc3d implements MutableLoc3d {
 
     @Override
     public double length() {
-        return (double) Math.sqrt(this.lengthSquared());
+        return Math.sqrt(this.lengthSquared());
     }
 
     @Override
